Extract chroot utils
